Fix lint

+++ contracts/pool/FarmingPool.sol
@@ -24,17 +24,31 @@ contract FarmingPool is SidePool {
 
   function _authorizeUpgrade(address newImplementation) internal virtual override onlyOwner {}
 
-
-  function stake(uint tokenType, uint256 lockupTime, uint256 tokenAmountOrID) external virtual override {
+  function stake(
+    uint256 tokenType,
+    uint256 lockupTime,
+    uint256 tokenAmountOrID
+  ) external virtual override {
     // mainIndex = type(uint16).max means no meanIndex
     require(tokenType == BLUEPRINT_STAKE_FOR_BOOST || tokenType == SEED_STAKE, "FarmingPool: unsupported token");
-    _stake(_msgSender(), tokenType, block.timestamp, block.timestamp.add(lockupTime * 1 days), type(uint16).max , tokenAmountOrID);
+    _stake(
+      _msgSender(),
+      tokenType,
+      block.timestamp,
+      block.timestamp.add(lockupTime * 1 days),
+      type(uint16).max,
+      tokenAmountOrID
+    );
   }
 
   function unstake(uint256 depositIndex) external override {
     Deposit memory deposit = users[_msgSender()].deposits[depositIndex];
-    require(deposit.tokenType == SEED_STAKE || deposit.tokenType == BLUEPRINT_STAKE_FOR_BOOST, "FarmingPool: invalid tokenType");
+    require(
+      deposit.tokenType == SEED_STAKE || deposit.tokenType == BLUEPRINT_STAKE_FOR_BOOST,
+      "FarmingPool: invalid tokenType"
+    );
     _unstakeDeposit(deposit);
   }
+
   uint256[50] private __gap;
 }
